% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_logit_lnlikelihood.R
\name{compute_logit_lnlikelihood}
\alias{compute_logit_lnlikelihood}
\title{Function that compute the logit log-likelihood function}
\usage{
compute_logit_lnlikelihood(param_constant, param_covariate, y, X)
}
\arguments{
\item{param_constant}{a parameter for the constant term}

\item{param_covariate}{parameters for covariates}

\item{y}{a vector of binary outcomes}

\item{X}{a covariate matrix}
}
\value{
logit log-likelihood function
}
\description{
Function that compute the logit log-likelihood function
}
